add Grphics (it need to create a structure for both devices and for queue familyes)

+++ src/main/App.cpp
@@ -52,7 +52,7 @@ VkInstance App::initVulkan() {
         .pNext = nullptr,
         .flags = {},
         .pApplicationInfo = &app_info,
-        .enabledLayerCount = layers.size(),
+        .enabledLayerCount = static_cast<uint32_t>(layers.size()),
         .ppEnabledLayerNames = layers.data(),
         .enabledExtensionCount = extensions_count,
         .ppEnabledExtensionNames = extensions,
